refactor: :recycle: CLI now on picocli
Updated CLI to now use the more modern and advanced picocli library
