--- cli/src/main/java/cli/Main.java
@@ -1,9 +1,0 @@
-package cli;
-
-public class Main {
-
-	public static void main(String[] args) {
-		ConsoleApplication consoleApplication = new ConsoleApplication();
-		consoleApplication.run(args);
-	}
-}
